+++ layers/BranchLayer.cu
@@ -12,8 +12,8 @@
 */
 
 __global__ void g_BranchLayer_backpropagation(
-	double** curDelta,
-	double* preDelta,
+	float** curDelta,
+	float* preDelta,
 	int curDeltaSize,
 	int len);
 
@@ -22,8 +22,8 @@ __global__ void g_BranchLayer_backpropagation(
  *dim3 thread= dim3(min(outputs[0]->getLen() / batch, 1024));
 */
 __global__ void g_BranchLayer_feedforward(
-	double* inputs,
-	double** outputs,
+	float* inputs,
+	float** outputs,
 	int len);
 
 
@@ -86,8 +86,8 @@ BranchLayer::BranchLayer(std::string name)
 
 	batch = Config::instance()->getBatchSize();
 	for(int i = 0; i < config->m_outputs.size(); i++){
-		outputs.push_back (new cuMatrix<double>(batch, outputDim * outputDim, outputAmount));
-		curDelta.push_back(new cuMatrix<double>(batch, outputDim * outputDim, outputAmount));
+		outputs.push_back (new cuMatrix<float>(batch, outputDim * outputDim, outputAmount));
+		curDelta.push_back(new cuMatrix<float>(batch, outputDim * outputDim, outputAmount));
 		mapId[config->m_outputs[i]] = i;
 	}
 
@@ -103,13 +103,13 @@ BranchLayer::BranchLayer(std::string name)
  *dim3 thread= dim3(min(outputs[0]->getLen() / batch, 1024));
 */
 __global__ void g_BranchLayer_feedforward(
-	double* inputs,
-	double** outputs,
+	float* inputs,
+	float** outputs,
 	int len)
 {
 	int branchId = blockIdx.x;
 
-	double* output = outputs[branchId];
+	float* output = outputs[branchId];
 
 	for(int i = 0; i < len; i += gridDim.y * blockDim.x){
 		int idx = i + threadIdx.x + blockIdx.y * blockDim.x;
@@ -125,15 +125,15 @@ __global__ void g_BranchLayer_feedforward(
 */
 
 __global__ void g_BranchLayer_backpropagation(
-	double** curDelta,
-	double* preDelta,
+	float** curDelta,
+	float* preDelta,
 	int curDeltaSize,
 	int len)
 {
 	for(int i = 0; i < len ; i += gridDim.x * blockDim.x){
 		int idx = i + blockDim.x * blockIdx.x + threadIdx.x;
 		if(idx < len){
-			double val = 0.0;
+			float val = 0.0;
 			for(int c = 0; c < curDeltaSize; c++){
 				val += curDelta[c][idx];
 			}
