+++ readData/readCIFAR100Data.h
@@ -12,8 +12,8 @@
 
 
 /*read trainning data and lables*/
-void read_CIFAR100_Data(cuMatrixVector<double> &trainX,
-	cuMatrixVector<double>&testX,
+void read_CIFAR100_Data(cuMatrixVector<float> &trainX,
+	cuMatrixVector<float>&testX,
 	cuMatrix<int>*&trainY,
 	cuMatrix<int>*&testY);
